# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/donations": {
      "post": {
        "operationId": "DonationsController_createDonation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDonationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Donations"
        ]
      },
      "get": {
        "operationId": "DonationsController_getAllDonations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Donations"
        ]
      }
    },
    "/api/donations/auth": {
      "post": {
        "operationId": "DonationsController_createAuthenticatedDonation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDonationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Donations"
        ]
      }
    },
    "/api/donations/confirm": {
      "post": {
        "operationId": "DonationsController_confirmPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Donations"
        ]
      }
    },
...
    "/api/donations/{id}": {
      "get": {
        "operationId": "DonationsController_getDonation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Donations"
        ]
      }
    },
...
    "/api/events": {
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evenement cree avec succes"
          },
          "400": {
            "description": "Donnees invalides"
          },
          "401": {
            "description": "Non authentifie"
          },
          "403": {
            "description": "Acces refuse"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Creer un nouvel evenement",
        "tags": [
          "Events"
        ]
      },
      "get": {
        "operationId": "EventsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Numero de page",
            "schema": {
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre par page",
            "schema": {
              "example": 10,
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Type d evenement",
            "schema": {
              "example": "workshop",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Statut",
            "schema": {
              "example": "published",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Recherche",
            "schema": {
              "example": "leadership",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste recuperee avec succes"
          }
        },
        "summary": "Liste des evenements",
        "tags": [
          "Events"
        ]
      }
    },
...
    "/api/events/stats": {
      "get": {
        "operationId": "EventsController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistiques recuperees"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Statistiques des evenements",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/upcoming": {
      "get": {
        "operationId": "EventsController_getUpcomingEvents",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre maximum",
            "schema": {
              "example": 5,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Evenements a venir",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID de l evenement",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evenement trouve"
          },
          "404": {
            "description": "Evenement non trouve"
          }
        },
        "summary": "Details d un evenement",
        "tags": [
          "Events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID de l evenement",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evenement mis a jour"
          },
          "404": {
            "description": "Evenement non trouve"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre a jour un evenement",
        "tags": [
          "Events"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID de l evenement",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evenement supprime"
          },
          "404": {
            "description": "Evenement non trouve"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un evenement",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/{id}/status": {
      "patch": {
        "operationId": "EventsController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID de l evenement",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Changer le statut d un evenement",
        "tags": [
          "Events"
        ]
      }
    },
...
  },
  "info": {
    "title": "Y-Mad API",
    "description": "\n      API complete du site Y-Mad (Youthful Madagascar)\n\n      INFORMATIONS DE CONTACT:\n      - Adresse: Carion, Antananarivo, Madagascar\n      - Telephone: +261 32 04 856 97\n      - Email: ymad.mg@gmail.com\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Y-Mad Association",
      "url": "https://y-mad.mg",
      "email": "ymad.mg@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Entrez votre token JWT"
      }
    },
    "schemas": {
...
      "CreateDonationDto": {
        "type": "object",
        "properties": {}
      },
      "ConfirmPaymentDto": {
        "type": "object",
        "properties": {}
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Titre de l'événement (français)"
          },
          "title_mg": {
            "type": "string",
            "description": "Titre de l'événement (malagasy)"
          },
          "description": {
            "type": "string",
            "description": "Description de l'événement (français)"
          },
          "description_mg": {
            "type": "string",
            "description": "Description de l'événement (malagasy)"
          },
          "type": {
            "type": "string",
            "description": "Type d'événement",
            "enum": [
              "camp",
              "workshop",
              "hackathon",
              "conference",
              "formation"
            ]
          },
          "status": {
            "type": "string",
            "description": "Statut de l'événement",
            "default": "draft"
          },
          "location": {
            "type": "string",
            "description": "Lieu de l'événement"
          },
          "region": {
            "type": "string",
            "description": "Région de l'événement"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "description": "Date de début"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "Date de fin"
          },
          "maxCapacity": {
            "type": "number",
            "description": "Capacité maximale"
          },
          "isFree": {
            "type": "boolean",
            "description": "Événement gratuit ?",
            "default": true
          },
          "price": {
            "type": "number",
            "description": "Prix en Ariary",
            "default": 0
          },
          "imageUrl": {
            "type": "string",
            "description": "URL de l'image"
          },
          "galleryImages": {
            "description": "Images de la galerie",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "program": {
            "type": "string",
            "description": "Programme détaillé"
          },
          "speakers": {
            "type": "string",
            "description": "Intervenants"
          }
        },
        "required": [
          "title",
          "description",
          "type",
          "location",
          "startDate",
          "isFree"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "title_mg": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "description_mg": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "startDate": {
            "format": "date-time",
            "type": "string"
          },
          "endDate": {
            "format": "date-time",
            "type": "string"
          },
          "maxCapacity": {
            "type": "number"
          },
          "isFree": {
            "type": "boolean"
          },
          "price": {
            "type": "number"
          },
          "imageUrl": {
            "type": "string"
          },
          "galleryImages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "program": {
            "type": "string"
          },
          "speakers": {
            "type": "string"
          }
        }
      },
...
    }
  }
}